start extracting config to copy file to workspace root
